add exception handling for cart add and checkout

+++ ecommerce/api/views.py
@@ -1,3 +1,4 @@
+from django.utils.ipv6 import ValidationError
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
@@ -14,8 +15,6 @@ def items_list(request):
     except Exception as error:
         return Response({"error": str(error)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
-
-
 @api_view(["GET", "POST"])
 def handle_cart_request(request):
     cart, created = Cart.objects.get_or_create(user=None)
@@ -30,9 +29,16 @@ def handle_cart_request(request):
     elif request.method == "POST":
         requested_item_id = request.data.get('item')
         item_quantity_requested = request.data.get('quantity')
+   
+        try:
+            item = Item.objects.get(id=requested_item_id)
+        except Item.DoesNotExist:
+            return Response({"error": "Item not found."}, status=status.HTTP_404_NOT_FOUND)
 
-        item = Item.objects.get(id = requested_item_id)
         quantity = int(item_quantity_requested)
+
+        if quantity > item.quantity:
+            return Response({"error": "Requested quantity exceeds available stock."}, status=status.HTTP_400_BAD_REQUEST)
 
         item.quantity -= quantity
         item.save()
@@ -45,14 +51,14 @@ def handle_cart_request(request):
         serializer = CartSerializer(added_cart_item)
 
         return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-
+   
 @api_view(["POST"])
 def cart_checkout(request):
     cart, created = Cart.objects.get_or_create(user=None)
     
-    cart_items = CartItem.objects.filter(cart=cart)
-
-    cart_items.delete()
-    
-    return Response({"message": "Checkout successful."}, status=status.HTTP_200_OK)
+    if not created:
+        cart_items = CartItem.objects.filter(cart=cart)
+        cart_items.delete()
+        return Response({"message": "Checkout successful."}, status=status.HTTP_200_OK)
+    else:
+        return Response({"error": "Cart is empty."}, status=status.HTTP_400_BAD_REQUEST)
